content/html/content/test/test_bug345624-2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=345624
-->
<head>
  <title>Test for Bug 345624</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=345624">Mozilla Bug 345624</a>
<p id="display"></p>
<div id="content" style="display: none">
  <input id='i'>
  <textarea id='t'></textarea>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 345624 **/

/**
 * This test is checking only tooLong related features
 * related to constraint validation.
 */

function checkTooLongValidity(element)
{
  element.value = "foo";
  ok(!element.validity.tooLong,
    "Element should not be too long when maxlength is not set");

  element.maxLength = 3;
  ok(!element.validity.tooLong,
    "Element should not be to long when maxlength = value length");

  element.maxLength = 5;
  ok(!element.validity.tooLong,
    "Element should not be too long when maxlength > value length");

  ok(element.validity.valid, "Element should be valid");

  element.maxLength = 2;
  ok(element.validity.tooLong,
    "Element should be too long when maxlength < value length");

  ok(!element.validity.valid,
    "Element should not be valid when it is too long");

  is(element.validationMessage,
    "The text is too long. It is 3 characters long and the limit is 2.",
    "The validation message text is not correct");
  ok(!element.checkValidity(), "The element should not be valid");
  element.setCustomValidity("custom message");
  is(element.validationMessage, "custom message",
    "Custom message should be shown instead of too long one");
}

checkTooLongValidity(document.getElementById('i'));
checkTooLongValidity(document.getElementById('t'));

</script>
</pre>
</body>
</html>
